Fix a layout issue.
This had been plaguing me for weeks. There was strange sizing
behaviour with codemirror that prevented me from letting the grid
container size it's items equally on the vertical axis.
Turns out the browser was getting weird ideas about what the minimum
height of the grid container should be, explicitly setting this to 0px
seems to have fixed that.
Now the website scales perfectly.